+++ entities/liquidity_provision.go
@@ -16,6 +16,9 @@ import (
 	"fmt"
 	"time"
 
+	"encoding/json"
+
+	v2 "code.vegaprotocol.io/protos/data-node/api/v2"
 	"code.vegaprotocol.io/protos/vega"
 	"github.com/jackc/pgtype"
 	"github.com/shopspring/decimal"
@@ -153,4 +156,39 @@ func (lp LiquidityProvision) ToRow() []interface{} {
 		lp.ID, lp.PartyID, lp.CreatedAt, lp.UpdatedAt, lp.MarketID,
 		lp.CommitmentAmount, lp.Fee, lp.Sells, lp.Buys, lp.Version,
 		lp.Status, lp.Reference, lp.VegaTime}
+}
+
+func (lp LiquidityProvision) Cursor() *Cursor {
+	lc := LiquidityProvisionCursor{
+		VegaTime: lp.VegaTime,
+		ID:       lp.ID.String(),
+	}
+	return NewCursor(lc.String())
+}
+
+func (lp LiquidityProvision) ToProtoEdge(_ ...any) *v2.LiquidityProvisionsEdge {
+	return &v2.LiquidityProvisionsEdge{
+		Node:   lp.ToProto(),
+		Cursor: lp.Cursor().Encode(),
+	}
+}
+
+type LiquidityProvisionCursor struct {
+	VegaTime time.Time `json:"vegaTime"`
+	ID       string    `json:"id"`
+}
+
+func (lc LiquidityProvisionCursor) String() string {
+	bs, err := json.Marshal(lc)
+	if err != nil {
+		panic(fmt.Errorf("could not marshal liquidity provision cursor: %w", err))
+	}
+	return string(bs)
+}
+
+func (lc *LiquidityProvisionCursor) Parse(cursorString string) error {
+	if cursorString == "" {
+		return nil
+	}
+	return json.Unmarshal([]byte(cursorString), lc)
 }