+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [733](https://github.com/vegaprotocol/data-node/issues/733) - Store chain info in database when using `SQL`
 - [748](https://github.com/vegaprotocol/data-node/issues/748) - Add REST endpoint to list `OracleData`
 - [761](https://github.com/vegaprotocol/data-node/issues/761) - Delete all badger stores, `SQL` stores only from now
+- [566](https://github.com/vegaprotocol/data-node/issues/566) - Liquidity provision pagination 
 
 ### 🐛 Fixes
 - [705](https://github.com/vegaprotocol/data-node/issues/705) - Market Depth returning incorrect book state